Beampipe is a bust, finally using plausible instead.
Beampipe has no average session duration and their documentation is very lacking as well as their site.
Can't find the data export even though it's claimed on the site. Support contact seems to be paid only. 

The setup was very straightforward... too straightforward. Took me a while to find the full Dashboard GUI 
(it turns out I had to click the weblink but that should have been in the docs).